Move handlers into handlers file
